src/backend/requirements.txt
Flask==2.3.2
pyjwt==2.8.0
datetime==5.2
uuid==1.30
Flask-SQLAlchemy==3.0.5
get_docker_secret==2.0.0
psycopg2-binary==2.9.7
